unsigned long

+++ 0x0B-malloc_free/100-argstostr.c
@@ -13,7 +13,7 @@
 char *argstostr(int ac, char **av)
 {
 	int i = 0;
-	int j = 0;
+	unsigned long int j = 0;
 	int len = 0;
 	int scl = 0;
 	char *s;
